move params and config to core + pass to flow ctx
