+++ consensus/core/src/config/mod.rs
@@ -1,0 +1,85 @@
+pub mod constants;
+pub mod params;
+
+use std::ops::Deref;
+
+use {
+    constants::perf::{PerfParams, PERF_PARAMS},
+    params::Params,
+};
+
+/// Various consensus configurations all bundled up under a single struct. Use `Config::new` for directly building from
+/// a `Params` instance. For anything more complex it is recommended to use `ConfigBuilder`. NOTE: this struct can be
+/// implicitly de-refed into `Params`
+#[derive(Clone)]
+pub struct Config {
+    /// Consensus params
+    pub params: Params,
+    /// Performance params
+    pub perf: PerfParams,
+
+    //
+    // Additional consensus configuration arguments which are not consensus sensitive
+    //
+    pub process_genesis: bool,
+    // TODO:
+    // is_archival: bool,
+    // enable_sanity_check_pruning_utxoset: bool,
+}
+
+impl Config {
+    pub fn new(params: Params) -> Self {
+        Self { params, perf: PERF_PARAMS, process_genesis: true }
+    }
+}
+
+impl AsRef<Params> for Config {
+    fn as_ref(&self) -> &Params {
+        &self.params
+    }
+}
+
+impl Deref for Config {
+    type Target = Params;
+
+    fn deref(&self) -> &Self::Target {
+        &self.params
+    }
+}
+
+pub struct ConfigBuilder {
+    config: Config,
+}
+
+impl ConfigBuilder {
+    pub fn new(params: Params) -> Self {
+        Self { config: Config::new(params) }
+    }
+
+    pub fn set_perf_params(mut self, perf: PerfParams) -> Self {
+        self.config.perf = perf;
+        self
+    }
+
+    pub fn edit_consensus_params<F>(mut self, edit_func: F) -> Self
+    where
+        F: Fn(&mut Params),
+    {
+        edit_func(&mut self.config.params);
+        self
+    }
+
+    pub fn skip_proof_of_work(mut self) -> Self {
+        self.config.params.skip_proof_of_work = true;
+        self
+    }
+
+    pub fn skip_adding_genesis(mut self) -> Self {
+        self.config.process_genesis = false;
+        self
+    }
+
+    pub fn build(self) -> Config {
+        self.config
+    }
+}
