+++ consensus/core/src/config/params.rs
@@ -1,0 +1,135 @@
+use crate::{BlockLevel, KType};
+use hashes::{Hash, HASH_SIZE};
+
+/// Consensus parameters. Contains settings and configurations which are consensus-sensitive.
+/// Changing one of these on a network node would exclude and prevent it from reaching consensus
+/// with the other unmodified nodes.
+#[derive(Clone)]
+pub struct Params {
+    pub genesis_hash: Hash,
+    pub ghostdag_k: KType,
+    pub timestamp_deviation_tolerance: u64,
+    pub target_time_per_block: u64,
+    pub max_block_parents: u8,
+    pub difficulty_window_size: usize,
+    pub genesis_timestamp: u64,
+    pub genesis_bits: u32,
+    pub mergeset_size_limit: u64,
+    pub merge_depth: u64,
+    pub finality_depth: u64,
+    pub pruning_depth: u64,
+    pub coinbase_payload_script_public_key_max_len: u8,
+    pub max_coinbase_payload_len: usize,
+    pub max_tx_inputs: usize,
+    pub max_tx_outputs: usize,
+    pub max_signature_script_len: usize,
+    pub max_script_public_key_len: usize,
+    pub mass_per_tx_byte: u64,
+    pub mass_per_script_pub_key_byte: u64,
+    pub mass_per_sig_op: u64,
+    pub max_block_mass: u64,
+    pub deflationary_phase_daa_score: u64,
+    pub pre_deflationary_phase_base_subsidy: u64,
+    pub coinbase_maturity: u64,
+    pub skip_proof_of_work: bool,
+    pub max_block_level: BlockLevel,
+    pub pruning_proof_m: u64,
+}
+
+impl Params {
+    pub fn expected_daa_window_duration_in_milliseconds(&self) -> u64 {
+        self.target_time_per_block * self.difficulty_window_size as u64
+    }
+}
+
+const DEFAULT_GHOSTDAG_K: KType = 18;
+pub const MAINNET_PARAMS: Params = Params {
+    genesis_hash: Hash::from_bytes([
+        0x58, 0xc2, 0xd4, 0x19, 0x9e, 0x21, 0xf9, 0x10, 0xd1, 0x57, 0x1d, 0x11, 0x49, 0x69, 0xce, 0xce, 0xf4, 0x8f, 0x9, 0xf9, 0x34,
+        0xd4, 0x2c, 0xcb, 0x6a, 0x28, 0x1a, 0x15, 0x86, 0x8f, 0x29, 0x99,
+    ]),
+    ghostdag_k: DEFAULT_GHOSTDAG_K,
+    timestamp_deviation_tolerance: 132,
+    target_time_per_block: 1000,
+    max_block_parents: 10,
+    difficulty_window_size: 2641,
+    genesis_timestamp: 1637609671037,
+    genesis_bits: 486722099,
+    mergeset_size_limit: (DEFAULT_GHOSTDAG_K as u64) * 10,
+    merge_depth: 3600,
+    finality_depth: 86400,
+    pruning_depth: 185798,
+    coinbase_payload_script_public_key_max_len: 150,
+    max_coinbase_payload_len: 204,
+
+    // This is technically a soft fork from the Go implementation since kaspad's consensus doesn't
+    // check these rules, but in practice it's enforced by the network layer that limits the message
+    // size to 1 GB.
+    // These values should be lowered to more reasonable amounts on the next planned HF/SF.
+    max_tx_inputs: 1_000_000_000,
+    max_tx_outputs: 1_000_000_000,
+    max_signature_script_len: 1_000_000_000,
+    max_script_public_key_len: 1_000_000_000,
+
+    mass_per_tx_byte: 1,
+    mass_per_script_pub_key_byte: 10,
+    mass_per_sig_op: 1000,
+    max_block_mass: 500_000,
+
+    // deflationary_phase_daa_score is the DAA score after which the pre-deflationary period
+    // switches to the deflationary period. This number is calculated as follows:
+    // We define a year as 365.25 days
+    // Half a year in seconds = 365.25 / 2 * 24 * 60 * 60 = 15778800
+    // The network was down for three days shortly after launch
+    // Three days in seconds = 3 * 24 * 60 * 60 = 259200
+    deflationary_phase_daa_score: 15778800 - 259200,
+    pre_deflationary_phase_base_subsidy: 50000000000,
+    coinbase_maturity: 100,
+    skip_proof_of_work: false,
+    max_block_level: 225,
+    pruning_proof_m: 1000,
+};
+
+pub const DEVNET_PARAMS: Params = Params {
+    genesis_hash: Hash::from_bytes([1u8; HASH_SIZE]), // TODO: Use golang devnet genesis here
+    ghostdag_k: DEFAULT_GHOSTDAG_K,
+    timestamp_deviation_tolerance: 132,
+    target_time_per_block: 1000,
+    max_block_parents: 10,
+    difficulty_window_size: 2641,
+    genesis_timestamp: 0,     // TODO: Use real value
+    genesis_bits: 0x1e21bc1c, // As observed on testnet
+    mergeset_size_limit: (DEFAULT_GHOSTDAG_K as u64) * 10,
+    merge_depth: 3600,
+    finality_depth: 86400,
+    pruning_depth: 185798,
+    coinbase_payload_script_public_key_max_len: 150,
+    max_coinbase_payload_len: 204,
+
+    // This is technically a soft fork from the Go implementation since kaspad's consensus doesn't
+    // check these rules, but in practice it's enforced by the network layer that limits the message
+    // size to 1 GB.
+    // These values should be lowered to more reasonable amounts on the next planned HF/SF.
+    max_tx_inputs: 1_000_000_000,
+    max_tx_outputs: 1_000_000_000,
+    max_signature_script_len: 1_000_000_000,
+    max_script_public_key_len: 1_000_000_000,
+
+    mass_per_tx_byte: 1,
+    mass_per_script_pub_key_byte: 10,
+    mass_per_sig_op: 1000,
+    max_block_mass: 500_000,
+
+    // deflationary_phase_daa_score is the DAA score after which the pre-deflationary period
+    // switches to the deflationary period. This number is calculated as follows:
+    // We define a year as 365.25 days
+    // Half a year in seconds = 365.25 / 2 * 24 * 60 * 60 = 15778800
+    // The network was down for three days shortly after launch
+    // Three days in seconds = 3 * 24 * 60 * 60 = 259200
+    deflationary_phase_daa_score: 15778800 - 259200,
+    pre_deflationary_phase_base_subsidy: 50000000000,
+    coinbase_maturity: 100,
+    skip_proof_of_work: false,
+    max_block_level: 250,
+    pruning_proof_m: 1000,
+};
