+++ kaspad/src/main.rs
@@ -176,7 +176,7 @@ pub fn main() {
     let rpc_core_server =
         Arc::new(RpcCoreServer::new(consensus.clone(), notify_service.notifier(), index_service.as_ref().map(|x| x.notifier())));
     let grpc_server = Arc::new(GrpcServer::new(grpc_server_addr, rpc_core_server.service()));
-    let p2p_service = Arc::new(P2pService::new(consensus.clone(), args.connect, args.listen));
+    let p2p_service = Arc::new(P2pService::new(consensus.clone(), &config, args.connect, args.listen));
 
     // TEMP: temp mining manager initialization just to make sure it complies with consensus
     let _mining_manager =
